+++ Tests/AsyncHTTPClientTests/HTTPClientTests+XCTest.swift
@@ -143,7 +143,7 @@ extension HTTPClientTests {
             ("testRequestWithHeaderTransferEncodingIdentityDoesNotFail", testRequestWithHeaderTransferEncodingIdentityDoesNotFail),
             ("testMassiveDownload", testMassiveDownload),
             ("testShutdownWithFutures", testShutdownWithFutures),
-            ("testMassiveHeader", testMassiveHeader),
+            ("testMassiveHeaderHTTP1", testMassiveHeaderHTTP1),
         ]
     }
 }
